requirements.txt
Flask==2.3.3
Flask-APScheduler==1.12.3
python-dotenv==1.0.0
Pillow==9.3.0
PyPDF2==3.0.0
